phase-II/backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.2.1
uvicorn==0.24.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0